src/main/java/com/jet/hexagonale/infrastruture/repository/cassandra/CassandraDbOrderRepository.java
package com.jet.hexagonale.infrastruture.repository.cassandra;

import java.util.Optional;
import java.util.UUID;

import com.jet.hexagonale.domain.Order;
import com.jet.hexagonale.domain.repository.OrderRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Component;
@Component
public class CassandraDbOrderRepository implements OrderRepository {

    private final SpringDataCassandraOrderRepository orderRepository;

    @Autowired
    public CassandraDbOrderRepository(SpringDataCassandraOrderRepository orderRepository) {
        this.orderRepository = orderRepository;
    }

    @Override
    public Optional<Order> findById(UUID id) {
        Optional<OrderEntity> orderEntity = orderRepository.findById(id);
        return orderEntity.map(OrderEntity::toOrder);
    }

    @Override
    public void save(Order order) {
        orderRepository.save(new OrderEntity(order));
    }

}
